+++ src/data/usecases/authentication/db-authentication.spec.ts
@@ -1,10 +1,5 @@
-import { AuthenticationModel } from '../../../domain/usecases/authentication'
-import { HashComparer } from '../../protocols/criptography/hash-compare'
-import { TokenGenerator } from '../../protocols/criptography/token-generator'
-import { LoadAccountByEmailRepository } from '../../protocols/db/load-account-by-email-repository'
-import { UpdateAccessTokenRepository } from '../../protocols/db/update-access-token-repository'
-import { AccountModel } from '../add-account /db-add-account-protocols'
 import { DbAuthentication } from './db-authentication'
+import { AccountModel, AuthenticationModel, LoadAccountByEmailRepository, UpdateAccessTokenRepository, HashComparer, TokenGenerator } from './db-authentication-protocols'
 
 const makeFakeAccount = (): AccountModel => ({
   id: 'any_id',
