+++ src/data/usecases/authentication/db-authentication.ts
@@ -1,9 +1,4 @@
-import { Authentication, AuthenticationModel } from '../../../domain/usecases/authentication'
-import { HashComparer } from '../../protocols/criptography/hash-compare'
-import { TokenGenerator } from '../../protocols/criptography/token-generator'
-import { LoadAccountByEmailRepository } from '../../protocols/db/load-account-by-email-repository'
-import { UpdateAccessTokenRepository } from '../../protocols/db/update-access-token-repository'
-
+import { Authentication, AuthenticationModel, LoadAccountByEmailRepository, UpdateAccessTokenRepository, HashComparer, TokenGenerator } from './db-authentication-protocols'
 export class DbAuthentication implements Authentication {
   private readonly loadAccountByEmailRepository: LoadAccountByEmailRepository
   private readonly hashCompare: HashComparer
